Move VibrationEffect to preview subpackage
