+++ sample/src/main/java/com/louiscad/splittiessample/preview/vibrator/VibrationEffect.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.louiscad.splittiessample.extensions
+package com.louiscad.splittiessample.preview.vibrator
 
 import android.Manifest
 import android.annotation.SuppressLint
